design(monorepo-2): spec + plan
